✨ feat 🐛 fix: sample api 추가, 엔티티 수정
회원가입 api추가
엔티티 수정 - diff 참고

+++ src/entities/Lenses.ts
@@ -24,7 +24,7 @@ export class Lenses {
   @Column('varchar', { name: 'focalLength', length: 100 })
   focalLength: string | null;
 
-  @Column('double', { name: 'aperture', precision: 22 })
+  @Column('float', { name: 'aperture' })
   aperture: number | null;
 
   @Column('enum', { name: 'type', enum: ['prime', 'zoom'] })
